requirements.txt
asgiref==3.8.1
Django==5.0.6
django-ckeditor==6.7.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-taggit==5.0.1
pillow==10.3.0
shortuuid==1.0.13
sqlparse==0.5.0
tzdata==2024.1
